Update scripts to use CLI pattern.
Refactor scripts to use CLI pattern:
  ./cloud-report generate ...
  ./cloud-report analyze ...

Make it possible to repeat --cloud-details flag to accept multiple JSON files.

Add go module definitions for the project and delete old/stale files.
