+++ go.mod
@@ -1,0 +1,12 @@
+module github.com/cockroachlabs/cloud-report
+
+go 1.13
+
+require (
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/spf13/cobra v1.0.0
+	github.com/spf13/viper v1.7.1
+	golang.org/x/net v0.0.0-20201002202402-0a1ea396d57c
+	golang.org/x/oauth2 v0.0.0-20200902213428-5d25da1a8d43
+	google.golang.org/api v0.32.0
+)
